requirements.txt
Click==7.0
Flask==1.0.3
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.16.3
Pillow==6.0.0
six==1.12.0
torch==1.1.0
torchvision==0.3.0
Werkzeug==0.15.4
